requirements.txt
scikit-learn
pandas
joblib
fastapi
uvicorn
streamlit
requests
dvc
mlflow
apache-airflow
certifi